refactor(dev): loader utils
