+++ packages/remix-dev/compiler/plugins/vanillaExtract/cached.ts
@@ -5,7 +5,7 @@ import type { Plugin } from "esbuild";
 
 import type { RemixConfig } from "../../../config";
 import type { CompileOptions } from "../../options";
-import { loaders } from "../../loaders";
+import { loaders } from "../../utils/loaders";
 import { getPostcssProcessor } from "../../utils/postcss";
 import { vanillaExtractSideEffectsPlugin } from "./sideEffects";
