+++ packages/remix-dev/compiler/serverjs/plugins/routes.ts
@@ -3,7 +3,7 @@ import * as fse from "fs-extra";
 import type esbuild from "esbuild";
 
 import type { RemixConfig } from "../../../config";
-import { getLoaderForFile } from "../../loaders";
+import { getLoaderForFile } from "../../utils/loaders";
 
 /**
  * This plugin loads route modules for the server build and prevents errors
